+++ modules/client/src/client-common/interfaces/plugin.ts
@@ -168,6 +168,7 @@ export type SubgraphVoterListItemBase = {
   voter: {
     address: string;
   };
+  voteReplaced: boolean
   voteOption: SubgraphVoteValues;
 };